Remove -Wno-tautological-compare from fusl build
For unsigned types, x < 0 comparisons are removed.

For enum tyes, x < 0 comparisions are replaced with (int)x < 0. C99 does
not specify the type of an enum. For e.g. ns_sect which has only
nonnegative values, it could be either signed or unsigned. In either
case, the cast to int should be safe.

R=viettrungluu@chromium.org

Review URL: https://codereview.chromium.org/1578383002 .

+++ fusl/src/network/ns_parse.c
@@ -105,7 +105,7 @@ int ns_parserr(ns_msg *handle, ns_sect section, int rrnum, ns_rr *rr)
 {
 	int r;
 
-	if (section < 0 || section >= ns_s_max) goto bad;
+	if ((int)section < 0 || section >= ns_s_max) goto bad;
 	if (section != handle->_sect) {
 		handle->_sect = section;
 		handle->_rrnum = 0;
@@ -168,4 +168,3 @@ int ns_name_uncompress(const unsigned char *msg, const unsigned char *eom,
 	if (r < 0) errno = EMSGSIZE;
 	return r;
 }
-
